:recycle: Drop HTTP delivery framework
replaced 'valyala/fasthttp' to native 'net/http' package, close #4

+++ internal/domain/code_challenge_method_test.go
@@ -114,7 +114,7 @@ func TestCodeChallengeMethod_String(t *testing.T) {
 func TestCodeChallengeMethod_Validate(t *testing.T) {
 	t.Parallel()
 
-	verifier, err := random.String(gofakeit.Number(43, 128))
+	verifier, err := random.String(uint8(gofakeit.Number(43, 128)))
 	if err != nil {
 		t.Fatalf("%+v", err)
 	}
